# Optimize top() and bottom() using an incremental aggregator