@jsternberg 2017-05-16
Benchmark comparison:

```
benchmark                    old ns/op     new ns/op     delta
BenchmarkSelect_Top_1K-8     381236352     64565949      -83.06%

benchmark                    old allocs     new allocs     delta
BenchmarkSelect_Top_1K-8     110            66             -40.00%

benchmark                    old bytes     new bytes     delta
BenchmarkSelect_Top_1K-8     529304706     12168         -100.00%
```

There should also be an improvement when using `SELECT top(value, host, <n>) ...`, but the code to benchmark this is less straightforward since the old implementation worked completely differently than this new implementation. I can try if needed, but it wouldn't be a 1-1 match between the code.